+++ src/components/text-field/text-field-unit/error-icon/index.ts
@@ -1,4 +1,4 @@
-import { makeStyleSheet } from "../../styles";
+import { makeStyleSheet } from "../../../styles";
 import styles from "./styles.css?inline";
 
 export class SpTextFieldErrorIcon extends HTMLElement {
