+++ src/index.ts
@@ -1,2 +1,3 @@
 export { SpButton } from "./components/button/sp-button";
-export { SpTextFieldUnit } from "./components/text-field-unit";
+export { SpTextFieldUnit } from "./components/text-field/text-field-unit";
+export { SpLabelUnit } from "./components/label-unit";
